+++ CHANGELOG.md
@@ -11,11 +11,16 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- [BREAKING]: Moved top level config options (TLS, Authentication, Authorization etc.) to new top level field `clusterConfig` ([#400]).
+- [BREAKING]: Support specifying Service type by moving `serviceType` (which was an experimental feature) to `clusterConfig.listenerClass`.
+  This enables us to later switch non-breaking to using `ListenerClasses` for the exposure of Services.
+  This change is breaking, because - for security reasons - we default to the `cluster-internal` `ListenerClass`.
+  If you need your cluster to be accessible from outside of Kubernetes you need to set `clusterConfig.listenerClass`
+  to `external-unstable` or `external-stable` ([#406]).
 - `operator-rs` `0.31.0` -> `0.39.0` ([#378], [#380], [#400], [#405]).
 - Bumped stackable image versions to `23.4.0-rc2` ([#378], [#380]).
 - Fragmented `TrinoConfig` ([#379]).
 - Enabled logging and log aggregation ([#380]).
-- [BREAKING]: Moved top level config options (TLS, Authentication, Authorization etc.) to new top level field `clusterConfig` ([#400]).
 
 ### Removed
 
@@ -27,6 +32,7 @@ All notable changes to this project will be documented in this file.
 [#391]: https://github.com/stackabletech/trino-operator/pull/391
 [#400]: https://github.com/stackabletech/trino-operator/pull/400
 [#405]: https://github.com/stackabletech/trino-operator/pull/405
+[#406]: https://github.com/stackabletech/trino-operator/pull/406
 
 ## [23.1.0] - 2023-01-23